cve-2015-3290: Bump up stack boundary requirement
When the input compiler defaults to AVX enabled, stack realignment
requirements causes gcc to fail to omit %rbp use, due to which the test
fails to clobber %rbp in inline asm.  Bump up the preferred stack
boundary to avoid this stack realignment for AVX, so the frame pointer
omission goes through.

Signed-off-by: Siddhesh Poyarekar <siddhesh@gotplt.org>
Reviewed-by: Martin Doucha <mdoucha@suse.cz>